Reorganizes ORC test suites
